Add registration to the TestUTTClientApi unit test.

+++ utt/utt-client-api/include/utt-client-api/ClientApi.hpp
@@ -29,7 +29,7 @@ namespace utt::client {
 struct ConfigInputParams {
   bool useBudget = true;  // Disable/enable usage of a budget token
   std::vector<std::string> participantsPublicKeys;
-  uint16_t threshold = 0;  // The number of participant shares required to reconstruct a signature
+  uint32_t threshold = 0;  // The number of participant shares required to reconstruct a signature
 };
 
 /// @brief Initialize the UTT Client API. Must be called before any other API function.
